#清洗模块  关联代码切换到mybatis
