+++ src/main/java/com/silita/biaodaa/rules/ruleImpl/hunan/HunanRelationRule.java
@@ -1,42 +1,65 @@
 package com.silita.biaodaa.rules.ruleImpl.hunan;
 
 import com.silita.biaodaa.rules.Interface.RelationRule;
+import com.silita.biaodaa.service.NoticeRelationService;
 import com.silita.biaodaa.utils.ComputeResemble;
 import com.silita.biaodaa.utils.MyStringUtils;
+import com.silita.biaodaa.utils.RuleUtils;
 import com.snatch.model.EsNotice;
 import org.apache.log4j.Logger;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.util.*;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
-/**
- * Created by dh on 2018/3/14.
- */
+import static com.silita.biaodaa.utils.RuleUtils.*;
+
 @Component
 public class HunanRelationRule extends HunanBaseRule implements RelationRule {
-    Logger logger = Logger.getLogger(this.getClass());
+    private Logger logger = Logger.getLogger(this.getClass());
+
+    @Autowired
+    private NoticeRelationService service;
+
+    /**
+     * 判断公告是否为张家界、长沙、邵阳、湘西、湘潭
+     * @param noticeUrl
+     * @return
+     */
+    public int especWebSite (String noticeUrl) {
+        for (int i = 0; i < normalUrl.length; i++) {
+            if (noticeUrl.contains(normalUrl[i])) {
+                return i;
+            }
+        }
+        return -1;
+    }
 
     @Override
     public Map<String, String> executeRule(EsNotice esNotice) {
         logger.info("湖南关联开始：[redisId:"+esNotice.getRedisId()+"][source:"+esNotice.getSource()+"][ur:"+esNotice.getUrl()+"]" + esNotice.getTitle() + esNotice.getOpenDate());
         Map<String,String> map = new HashMap<String,String>();
         List<String> relationList = new ArrayList<String>(); // 关联列表
-        int urlIndex = urlIndexOf(esNotice.getUrl());
-        if (urlIndex != -1) {
+        int especIdx = especWebSite(esNotice.getUrl());
+        if (especIdx != -1) {
+            Map argMap = new HashMap<>();
             logger.info("#####  启用张家界、邵阳、湘西、湘潭、长沙 公告关联规则！  #####");
-            // 张家界，邵阳，湘西，湘潭，长沙 公告关联规则
             String noticeUrl = esNotice.getUrl();
-            String source = esNotice.getSource();
+            argMap.put("source",esNotice.getSource());
             String urlKey = "";
-            if (urlIndex < 3) {
+            if (especIdx < 3) {
                 // 张家界、邵阳、湘西
                 urlKey = noticeUrl.substring(noticeUrl.indexOf("&tpid=") + 6);
-                relationList = snatchNoticeHuNanDao.querysLikeUrl(urlKey,source);
-            } else if (urlIndex == 3) {
+                argMap.put("url","%"+urlKey+"%");
+                relationList = service.querysLikeUrl(argMap);
+            } else if (especIdx == 3) {
                 // 湘潭
                 urlKey = noticeUrl.substring(0,noticeUrl.lastIndexOf("?") + 1);
-                relationList = snatchNoticeHuNanDao.querysLikeUrl(urlKey,source);
-            } else if (urlIndex == 4) {
+                argMap.put("url","%"+urlKey+"%");
+                relationList = service.querysLikeUrl(argMap);
+            } else if (especIdx == 4) {
                 // 长沙
                 String relationUlrs = "";
                 if (esNotice.getDetail() != null && MyStringUtils.isNotNull(esNotice.getDetail().getRelationUrl())) {
@@ -46,7 +69,8 @@ public class HunanRelationRule extends HunanBaseRule implements RelationRule {
                 }
                 if (relationUlrs.split(",").length > 1) {
                     String[] relation_urls = relationUlrs.split(",");
-                    relationList = snatchNoticeHuNanDao.querysLikeUrl(Arrays.asList(relation_urls),source);
+                    argMap.put("urls",Arrays.asList(relation_urls));
+                    relationList = service.querysLikeUrl(argMap);
                 }
             }
         } else {
@@ -60,14 +84,13 @@ public class HunanRelationRule extends HunanBaseRule implements RelationRule {
             tempTitle = replaceStrSymbol(tempTitle); // 替换符号空格为%，标题前后添加%
 //            logger.info("####  处理后的模糊匹配词：" + tempTitle + "  ####");
             String websiteUrl = MyStringUtils.parseWebSiteUrl(esNotice.getUrl());
-            List<Map<String,Object>> searchResult = snatchNoticeHuNanDao.querySimilarityNotice(esNotice,websiteUrl,tempTitle);
-            // 相似度搜索公告
-            List<Map<String,Object>> result2 = snatchNoticeHuNanDao.querySimilarityNotice(esNotice,websiteUrl,null);
+            List<Map<String,Object>> searchResult = service.querySimilarityNotice(esNotice.getOpenDate(),websiteUrl,tempTitle);
 
-            // set去重
             Set<Map<String,Object>> set = new HashSet<Map<String,Object>>(searchResult);
 
             String title2 = clearKeyWord(title);
+            // 增加相似度匹配的公告
+            List<Map<String,Object>> result2 = service.querySimilarityNotice(esNotice.getOpenDate(),websiteUrl,null);
             for (Map<String,Object> mp : result2) { // 只保留与新进公告标题95%相似度以上的
                 String hstyTitle = clearKeyWord(String.valueOf(mp.get("title")));
                 if (ComputeResemble.similarDegreeWrapper(title2, hstyTitle) > 0.95) {
@@ -76,9 +99,7 @@ public class HunanRelationRule extends HunanBaseRule implements RelationRule {
             }
 
             searchResult = new ArrayList<Map<String,Object>>(set);
-
 //            logger.info("#### 模糊匹配  .. resultSize：" + searchResult.size() + "  ####");
-
             // 数据过滤
             if (!searchResult.isEmpty() && searchResult.size() > 1) { // size为 1 ，无相关公告
                 searchResult = noticeFilter(searchResult,esNotice);
@@ -117,5 +138,126 @@ public class HunanRelationRule extends HunanBaseRule implements RelationRule {
         }
         logger.info("#####  湖南关联结束！  #####[redisId:"+esNotice.getRedisId()+"][source:"+esNotice.getSource()+"][ur:"+esNotice.getUrl()+"]" + esNotice.getTitle() + esNotice.getOpenDate());
         return map;
+    }
+
+    /**
+     * 公告过滤
+     * @param searchResult
+     * @param notice
+     * @return
+     */
+    public List<Map<String,Object>> noticeFilter (List<Map<String,Object>> searchResult,EsNotice notice) {
+        // 数据过滤
+        logger.info("####  数据过滤 .. resultSize：" + searchResult.size() + "  ####");
+        String title = notice.getTitle();
+
+        // 标段过滤
+        if (searchResult.size() > 1) {
+            Iterator<Map<String,Object>> it = searchResult.iterator();
+            while (it.hasNext()){
+                String resultTitle = String.valueOf(it.next().get("title"));
+                if (title.contains("标段")) {
+                    if (!resultTitle.contains("标段")) {
+                        it.remove();
+                    }else {
+                        // 取出两个标题中的数字或英文
+                        String titleNumStr = null;
+                        if (title.lastIndexOf("标段") != -1) {
+                            titleNumStr = getNumStr(title.substring(0, title.lastIndexOf("标段")));
+                        } else {
+                            titleNumStr = getNumStr(title);
+                        }
+                        String resultTitleNumStr = null;
+                        if (resultTitle.lastIndexOf("标段") != -1) {
+                            resultTitleNumStr = getNumStr(resultTitle.substring(0, resultTitle.lastIndexOf("标段")));
+                        } else {
+                            resultTitleNumStr = getNumStr(resultTitle);
+                        }
+                        if (!titleNumStr.equals(resultTitleNumStr)) {
+                            // 俩个标题的数字或英文不一致
+                            it.remove();
+                        }
+                    }
+                } else if (resultTitle.contains("标段")) {
+                    // 新进公告没有标段，相关公告有标段
+                    it.remove();
+                }
+                logger.info("####  标段过滤 ..  resultSize：" + searchResult.size() + "  ####");
+            }
+        }
+
+        // 公告类型过滤
+        if (searchResult.size() > 1) {
+            Iterator<Map<String,Object>> it = searchResult.iterator();
+            while (it.hasNext()) {
+                String resultTitle = String.valueOf(it.next().get("title"));
+                if (RuleUtils.keyWords3IndexOf(resultTitle,keyWords3) != RuleUtils.keyWords3IndexOf(title,keyWords3)) {
+                    it.remove();
+                }
+            }
+            logger.info("####  公告类型过滤 ..  resultSize：" + searchResult.size() + "  ####");
+        }
+
+        // 项目次数过滤
+        if (searchResult.size() > 1) {
+            Iterator<Map<String,Object>> it = searchResult.iterator();
+            String regex = "(第).{1}?(次|批|包)";
+            Pattern pa = Pattern.compile(regex);
+            while (it.hasNext()) {
+                String resultTitle = String.valueOf(it.next().get("title"));
+                Matcher ma = pa.matcher(title);
+                if (ma.find()) {
+                    String sabi = ma.group();
+                    ma = pa.matcher(resultTitle);
+                    if (ma.find()) {
+                        // 新公告与历史公告都存在第几次字段
+                        int titleRegIndex = title.indexOf(sabi);
+                        int titleKeyIndex = keyWordsIndex(title, keyWords4);
+                        int historyTitleRegIndex = resultTitle.indexOf(sabi);
+                        int histotyTitleKeyIndex = keyWordsIndex(resultTitle, keyWords4);
+                        if (titleKeyIndex == -1 || titleKeyIndex > titleRegIndex) {
+                            // 新进公告没有关键字或关键字在相关字段前
+                            if (histotyTitleKeyIndex == -1 || histotyTitleKeyIndex > historyTitleRegIndex) {
+                            } else {it.remove();}
+                        } else {
+                            if (histotyTitleKeyIndex == -1 || histotyTitleKeyIndex > historyTitleRegIndex) {
+                                it.remove();
+                            }
+                        }
+                    } else {
+                        // 新公告存在第几次字段，历史公告无第几次字段,过滤掉
+                        it.remove();
+                    }
+                } else {
+                    // 新进公告无第几次字段，历史公告存在第几次字段，过滤掉
+                    ma = pa.matcher(resultTitle);
+                    if (ma.find()) {
+                        it.remove();
+                    }
+                }
+            }
+            logger.info("####  项目次数过滤 ..  historyNotices：" + searchResult.size() + "  ####");
+        }
+
+        // 括号内容过滤
+        if (searchResult.size() > 1) {
+            if (contaninsBracket(title)) {
+                int keyIndex = keyWordsIndex(title,keyWords5); // 获取第一个关键字的位置
+                if (keyIndex != -1) {
+                    String tempTitle = title.substring(0,keyIndex);
+                    if (contaninsBracket(tempTitle)) {
+                        Iterator<Map<String,Object>> it = searchResult.iterator();
+                        while (it.hasNext()) {
+                            String resultTitle = String.valueOf(it.next().get("title"));
+                            if (!compareBracketStr(tempTitle,resultTitle,keyWords6)) {
+                                it.remove();
+                            }
+                        }
+                    }
+                }
+            }
+            logger.info("####  括号内容过滤 .. resultSize: " + searchResult.size() + "  ####");
+        }
+        return searchResult;
     }
 }
